+++ crates/core/hadris-storage/src/async.rs
@@ -8,6 +8,8 @@ macro_rules! io_transform {
     ($($item:tt)*) => { $($item)* };
 }
 
+use hadris_io::r#async::MaybeSend;
+
 #[allow(clippy::duplicate_mod)]
 #[path = "api.rs"]
 mod api;